# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreatePlatformRequest:
    properties:
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - BANK
        - E_WALLET
        - CASH
        - ATM
        type: string
    required:
    - name
    - type
    type: object
...
  dto.UpdatePlatformRequest:
    properties:
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - BANK
        - E_WALLET
        - CASH
        - ATM
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive financial management system with smart allocation engine,
    transaction tracking, and detailed reports
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coin Backend API
  version: "1.0"
paths:
...
  /v1/platforms:
    get:
      consumes:
      - application/json
      description: Get a list of all platforms with pagination
      parameters:
      - description: 'Limit (default: 10, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Skip (default: 0)'
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Platforms retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all platforms
      tags:
      - Platforms
    post:
      consumes:
      - application/json
      description: Create a new platform (admin only)
      parameters:
      - description: Platform details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlatformRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Platform created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new platform
      tags:
      - Platforms
  /v1/platforms/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a platform (admin only)
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Platform deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Platform not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete platform
      tags:
      - Platforms
    get:
      consumes:
      - application/json
      description: Get a specific platform by ID
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Platform retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid platform ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Platform not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get platform by ID
      tags:
      - Platforms
    put:
      consumes:
      - application/json
      description: Update a platform (admin only)
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      - description: Platform update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Platform updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Platform not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update platform
      tags:
      - Platforms
  /v1/platforms/active:
    get:
      consumes:
      - application/json
      description: Get a list of active platforms with pagination
      parameters:
      - description: 'Limit (default: 10, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Skip (default: 0)'
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Active platforms retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List active platforms
      tags:
      - Platforms
  /v1/platforms/type/{type}:
    get:
      consumes:
      - application/json
      description: Get a list of platforms filtered by type with pagination
      parameters:
      - description: Platform type (BANK, E_WALLET, CASH, ATM)
        in: query
        name: type
        required: true
        type: string
      - description: 'Limit (default: 10, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Skip (default: 0)'
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Platforms retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List platforms by type
      tags:
      - Platforms
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"